Add a new class! macro that caches classes
This macro should improve performance and ergonomics. Like the sel!
macro, it will cache the gotten class so that repeated calls need not
call `objc_getClass`.

The macro only accepts an identifier (not a string), so the class being
looked up is static, and therefore it seems reasonable for this macro to
panic if the requested class does not exist. This improves ergonomics
because users need not call `unwrap` on the result (like was usually
done with `Class::get`).

This fixes #65.