Stop dereferencing msg_send!'s first argument
In Objective-C, having a null pointer receiver is valid, but with `msg_send!` it would be converted into a reference, which is undefined behavior.

Additionally, there might be a mutable reference somewhere else in the program, which would now be aliased.

Note: The reference is converted into a pointer immediately after, but it is still UB.