Make Sel an opaque struct like Ivar, Method, Class, Protocol and Object
The discrepancy between usage of `Class`/`Object`/... and `Sel` annoyed me, the others were used behind pointers / references, while `Sel` was consumed directly.

This changes the definition of `Sel` to be like the other types (so it has to be used as `&Sel`), and also opens up for the possiblity of non-static selectors (if that ever becomes a thing).

This is a breaking change, but fortunately most cases where `Sel` is used in user code (e.g. `add_method`) the error is caught at compile-time, with recommendations on how to fix it.